Use consistent variable type

+++ gnubg/randomorg.c
@@ -38,7 +38,7 @@ RandomOrgCallBack(void *pvRawData, size_t nSize, size_t nNumMemb, void *pvUserDa
     size_t nNewDataLen = nSize * nNumMemb;
     RandomData *pvRandomData = (RandomData *) pvUserData;
     unsigned int i;
-    unsigned int iNumRead = pvRandomData->nNumRolls;
+    size_t iNumRead = pvRandomData->nNumRolls;
     char *szRawData = (char *) pvRawData;
 
 #if defined(RANDOMORG_DEBUG)
